Completely refactor gameplay
Gameplay is completely refactored. The design is based on the
article *Build a House from Inside Out* and seems to make more
sense now. I need to check if the game is over and handle that,
then move to the dom manipualtion.

One future note is to use button instead of divs for the cells
the user can click on for accessibility reasons.